breaking: Pool fee is always expressed in hundredths of basis point
Previously we aimed to express it in BIPs outside of the SDK, but it resulted in chaos - in some places it was in BIPs, in others it was in 1/100 BIP.

Now pool fee is always is in 1/100 BIPs.

NOTE: Only pool fee is in 1/100 BIPs. midPriceDelta is still in BIPs.